Updated disable/enable mic btn api and endcall
