+++ template/src/components/ChatContext.ts
@@ -66,8 +66,6 @@ export enum controlMessageEnum {
   requestAudio = '7',
   //newUserJoined = '8',
   kickScreenshare = '9',
-  disableButton = '10',
-  endCallForEveryone = '11',
 }
 
 const ChatContext = createContext(null as unknown as chatContext);
